transformers_env_requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
abseil-cpp=20211102.0=hd77b12b_0
accelerate=0.21.0=pyhd8ed1ab_0
aiohttp=3.8.3=py310h2bbff1b_0
aiosignal=1.2.0=pyhd3eb1b0_0
appdirs=1.4.4=pyhd3eb1b0_0
arrow-cpp=11.0.0=ha81ea56_2
asttokens=2.0.5=pyhd3eb1b0_0
async-timeout=4.0.2=py310haa95532_0
attrs=22.1.0=py310haa95532_0
aws-c-common=0.6.8=h2bbff1b_1
aws-c-event-stream=0.1.6=hd77b12b_6
aws-checksums=0.1.11=h2bbff1b_2
aws-sdk-cpp=1.8.185=hd77b12b_1
backcall=0.2.0=pyhd3eb1b0_0
bitsandbytes=0.43.1=pypi_0
blas=1.0=mkl
boost-cpp=1.82.0=h59b6b97_1
bottleneck=1.3.5=py310h9128911_0
brotli=1.0.9=h2bbff1b_7
brotli-bin=1.0.9=h2bbff1b_7
brotlipy=0.7.0=py310h2bbff1b_1002
bzip2=1.0.8=he774522_0
c-ares=1.19.0=h2bbff1b_0
ca-certificates=2023.08.22=haa95532_0
certifi=2023.7.22=py310haa95532_0
cffi=1.15.1=py310h2bbff1b_3
charset-normalizer=2.0.4=pyhd3eb1b0_0
click=8.0.4=py310haa95532_0
colorama=0.4.6=py310haa95532_0
comm=0.1.2=py310haa95532_0
contourpy=1.0.5=py310h59b6b97_0
cryptography=41.0.2=py310hac1b9e3_0
cuda=11.7.1=0
cuda-cccl=11.7.91=0
cuda-command-line-tools=11.7.1=0
cuda-compiler=11.7.1=0
cuda-cudart=11.7.99=0
cuda-cudart-dev=11.7.99=0
cuda-cuobjdump=11.7.91=0
cuda-cupti=11.7.101=0
cuda-cuxxfilt=11.7.91=0
cuda-demo-suite=12.0.76=0
cuda-documentation=12.0.76=0
cuda-libraries=11.7.1=0
cuda-libraries-dev=11.7.1=0
cuda-memcheck=11.8.86=0
cuda-nsight-compute=12.0.0=0
cuda-nvcc=11.7.99=0
cuda-nvdisasm=12.0.76=h63175ca_0
cuda-nvml-dev=11.7.91=0
cuda-nvprof=12.0.90=h63175ca_0
cuda-nvprune=11.7.91=0
cuda-nvrtc=11.7.99=0
cuda-nvrtc-dev=11.7.99=0
cuda-nvtx=11.7.91=0
cuda-nvvp=12.0.90=h63175ca_0
cuda-runtime=11.7.1=0
cuda-sanitizer-api=12.0.90=h63175ca_0
cuda-toolkit=11.7.1=0
cuda-tools=11.7.1=0
cuda-version=12.0=hffde075_2
cuda-visual-tools=11.7.1=0
cycler=0.11.0=pyhd3eb1b0_0
daal4py=2023.1.1=py310hf497b98_0
dal=2023.1.1=h59b6b97_48681
dataclasses=0.8=pyh6d0b6a4_7
datasets=2.19.0=pypi_0
debugpy=1.6.7=py310hd77b12b_0
decorator=5.1.1=pyhd3eb1b0_0
dill=0.3.6=py310haa95532_0
docker-pycreds=0.4.0=pyhd3eb1b0_0
docstring-parser=0.16=pypi_0
einops=0.6.1=pypi_0
et_xmlfile=1.1.0=py310haa95532_0
evaluate=0.4.1=pypi_0
executing=0.8.3=pyhd3eb1b0_0
filelock=3.9.0=py310haa95532_0
fonttools=4.25.0=pyhd3eb1b0_0
freetype=2.12.1=ha860e81_0
frozenlist=1.3.3=py310h2bbff1b_0
fsspec=2024.3.1=pypi_0
gflags=2.2.2=ha925a31_0
giflib=5.2.1=h8cc25b3_3
gitdb=4.0.7=pyhd3eb1b0_0
gitpython=3.1.30=py310haa95532_0
glog=0.5.0=hd77b12b_0
grpc-cpp=1.48.2=hfe90ff0_1
huggingface-cli=0.1=pypi_0
huggingface-hub=0.22.2=pypi_0
icc_rt=2022.1.0=h6049295_2
idna=3.4=py310haa95532_0
ijson=3.2.0.post0=pypi_0
importlib-metadata=6.0.0=py310haa95532_0
importlib_metadata=6.0.0=hd3eb1b0_0
intel-openmp=2023.1.0=h59b6b97_46319
ipykernel=6.19.2=py310h9909e9c_0
ipython=8.12.2=py310haa95532_0
ipywidgets=8.1.0=pypi_0
jedi=0.18.1=py310haa95532_1
joblib=1.2.0=py310haa95532_0
jpeg=9e=h2bbff1b_1
jupyter_client=8.1.0=py310haa95532_0
jupyter_core=5.3.0=py310haa95532_0
jupyterlab-widgets=3.0.8=pypi_0
kiwisolver=1.4.4=py310hd77b12b_0
lerc=3.0=hd77b12b_0
libboost=1.82.0=hae598e9_1
libbrotlicommon=1.0.9=h2bbff1b_7
libbrotlidec=1.0.9=h2bbff1b_7
libbrotlienc=1.0.9=h2bbff1b_7
libcublas=11.10.3.66=0
libcublas-dev=11.10.3.66=0
libcufft=10.7.2.124=0
libcufft-dev=10.7.2.124=0
libcurand=10.3.1.50=h63175ca_0
libcurand-dev=10.3.1.50=h63175ca_0
libcurl=8.1.1=h86230a5_0
libcusolver=11.4.0.1=0
libcusolver-dev=11.4.0.1=0
libcusparse=11.7.4.91=0
libcusparse-dev=11.7.4.91=0
libdeflate=1.17=h2bbff1b_0
libevent=2.1.12=h56d1f94_1
libffi=3.4.4=hd77b12b_0
libnpp=11.7.4.75=0
libnpp-dev=11.7.4.75=0
libnvjpeg=11.8.0.2=0
libnvjpeg-dev=11.8.0.2=0
libpng=1.6.39=h8cc25b3_0
libprotobuf=3.20.3=h23ce68f_0
libsodium=1.0.18=h62dcd97_0
libssh2=1.10.0=he2ea4bf_2
libthrift=0.15.0=h4364b78_2
libtiff=4.5.0=h6c2663c_2
libuv=1.44.2=h2bbff1b_0
libwebp=1.2.4=hbc33d0d_1
libwebp-base=1.2.4=h2bbff1b_1
lz4-c=1.9.4=h2bbff1b_0
markdown-it-py=3.0.0=pypi_0
matplotlib-base=3.7.1=py310h4ed8f06_1
matplotlib-inline=0.1.6=py310haa95532_0
mdurl=0.1.2=pypi_0
mkl=2023.1.0=h6b88ed4_46357
mkl-service=2.4.0=py310h2bbff1b_1
mkl_fft=1.3.6=py310h4ed8f06_1
mkl_random=1.2.2=py310h4ed8f06_1
multidict=6.0.2=py310h2bbff1b_0
multiprocess=0.70.14=py310haa95532_0
munkres=1.1.4=py_0
nest-asyncio=1.5.6=py310haa95532_0
nltk=3.8.1=pypi_0
nsight-compute=2022.4.0.15=0
numexpr=2.8.4=py310h2cd9be0_1
numpy=1.25.2=py310h055cbcc_0
numpy-base=1.25.2=py310h65a83cf_0
openai=0.27.2=pyhd8ed1ab_0
openpyxl=3.0.10=py310h2bbff1b_0
openssl=3.0.10=h2bbff1b_2
orc=1.7.4=h623e30f_1
packaging=23.0=py310haa95532_0
pandas=2.0.3=py310h4ed8f06_0
pandas-stubs=1.5.3.230203=py310haa95532_0
parso=0.8.3=pyhd3eb1b0_0
pathtools=0.1.2=pyhd3eb1b0_1
peft=0.10.0=pypi_0
pickleshare=0.7.5=pyhd3eb1b0_1003
pillow=9.4.0=py310hd77b12b_0
pip=23.2.1=py310haa95532_0
platformdirs=3.10.0=py310haa95532_0
plotly=5.9.0=py310haa95532_0
prompt-toolkit=3.0.36=py310haa95532_0
protobuf=3.20.3=py310hd77b12b_0
psutil=5.9.0=py310h2bbff1b_0
pure_eval=0.2.2=pyhd3eb1b0_0
pyarrow=16.0.0=pypi_0
pyarrow-hotfix=0.6=pypi_0
pycparser=2.21=pyhd3eb1b0_0
pygments=2.15.1=py310haa95532_1
pyopenssl=23.2.0=py310haa95532_0
pyparsing=3.0.9=py310haa95532_0
pysocks=1.7.1=py310haa95532_0
python=3.10.12=he1021f5_0
python-dateutil=2.8.2=pyhd3eb1b0_0
python-tzdata=2023.3=pyhd3eb1b0_0
python-xxhash=2.0.2=py310h2bbff1b_1
pytorch=1.13.1=py3.10_cuda11.7_cudnn8_0
pytorch-cuda=11.7=h67b0de4_1
pytorch-mutex=1.0=cuda
pytz=2022.7=py310haa95532_0
pywin32=305=py310h2bbff1b_0
pyyaml=6.0=py310h2bbff1b_1
pyzmq=25.1.0=py310hd77b12b_0
re2=2022.04.01=hd77b12b_0
regex=2022.7.9=py310h2bbff1b_0
requests=2.31.0=py310haa95532_0
responses=0.13.3=pyhd3eb1b0_0
rich=13.7.1=pypi_0
rouge=1.0.1=pypi_0
sacremoses=0.0.43=pyhd3eb1b0_0
safetensors=0.3.2=py310h062c2fa_0
scikit-learn=1.3.0=py310h4ed8f06_0
scikit-learn-intelex=2023.1.1=py310haa95532_0
scipy=1.11.1=py310h309d312_0
sentencepiece=0.1.99=py310h59b6b97_0
sentry-sdk=1.9.0=py310haa95532_0
setproctitle=1.2.2=py310h2bbff1b_0
setuptools=68.0.0=py310haa95532_0
shtab=1.7.1=pypi_0
six=1.16.0=pyhd3eb1b0_1
smmap=4.0.0=pyhd3eb1b0_0
snappy=1.1.9=h6c2663c_0
sqlite=3.41.2=h2bbff1b_0
stack_data=0.2.0=pyhd3eb1b0_0
tbb=2021.8.0=h59b6b97_0
tenacity=8.2.2=py310haa95532_0
threadpoolctl=2.2.0=pyh0d69192_0
tk=8.6.12=h2bbff1b_0
tokenizer=3.4.3=pypi_0
tokenizers=0.13.2=py310h49fca51_1
torchaudio=0.13.1=pypi_0
torchmetrics=0.11.4=py310h9909e9c_1
torchvision=0.15.2=cpu_py310h7187fe4_0
tornado=6.3.2=py310h2bbff1b_0
tqdm=4.65.0=py310h9909e9c_0
traitlets=5.7.1=py310haa95532_0
transformers=4.31.0=pyhd8ed1ab_0
trl=0.8.6=pypi_0
types-pytz=2022.4.0.0=py310haa95532_1
typing-extensions=4.7.1=py310haa95532_0
typing_extensions=4.7.1=py310haa95532_0
tyro=0.8.3=pypi_0
tzdata=2023c=h04d1e81_0
ucrt=10.0.20348.0=haa95532_0
urllib3=1.26.16=py310haa95532_0
utf8proc=2.6.1=h2bbff1b_0
vc=14.2=h21ff451_1
vc14_runtime=14.36.32532=hfdfe4a8_17
vs2015_runtime=14.36.32532=h05e6639_17
wandb=0.15.8=pyhd8ed1ab_0
wcwidth=0.2.5=pyhd3eb1b0_0
wheel=0.38.4=py310haa95532_0
widgetsnbextension=4.0.8=pypi_0
win_inet_pton=1.1.0=py310haa95532_0
xxhash=0.8.0=h2bbff1b_3
xz=5.4.2=h8cc25b3_0
yaml=0.2.5=he774522_0
yarl=1.8.1=py310h2bbff1b_0
zeromq=4.3.4=hd77b12b_0
zipp=3.11.0=py310haa95532_0
zlib=1.2.13=h8cc25b3_0
zstd=1.5.5=hd43e919_0
